home page now displays the user's picture.  Reorganized file structure.  Added global variables for path.  Minimal styling.  Still need to work on uploading the file and editing other profile information
