+++ App/Utils/filepath.php
@@ -1,18 +1,27 @@
 <?php
+
 $doc_root = filter_input(INPUT_SERVER, 'DOCUMENT_ROOT');
 
 $uri = filter_input(INPUT_SERVER, 'REQUEST_URI');
 $dirs = explode('/', $uri);
-$app_path = '';
+$app_path = '/';
 
-$i = 0;
 foreach($dirs as $directory) {
-    if ($i == 0) {
-        $app_path += '/' . $directory . '/';
-    } else {
-        $app_path += $directory . '/';
+	if (!empty($directory)) {
+        $app_path .= $directory . '/';
     }
-    $i++;
+}
+
+$picture_path = '/';
+foreach($dirs as $directory) {
+    if (!empty($directory)) {
+    	if ($directory == 'App') {
+    		$picture_path .= 'ProfilePictures/';
+    		break;
+    	}
+    	else
+        	$picture_path .= $directory . '/';
+    }
 }
 
 set_include_path($doc_root . $app_path);
